working without compute shader for every dds format

+++ vvvv_plugin_dev/VVVV.Packs.Image/src/nodes/plugins/Image/DX11Player/Native/Channel.h
@@ -26,7 +26,7 @@ public:
         return m_queue.empty();
     }
 
-    int size() const
+    size_t size() const
     {
 		std::lock_guard<std::mutex> lock(m_mutex);
         return m_queue.size();
